requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-pinecone>=0.0.1
langgraph>=0.0.20
pinecone-client>=2.2.4
python-docx>=1.0.1
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
streamlit>=1.28.0
tiktoken>=0.5.1
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10
docx2txt>=0.8 